Update Email_Login
This is the framework for the authentication system. It works but it needs to be edited to deal with exceptions.  I tried to comment on the places that need some work. The account class is at the bottom of this but it probably needs to be separated into a new file when editing and running.